+++ msg.go
@@ -2,6 +2,7 @@ package main
 
 import (
 	"fmt"
+	"log"
 
 	"github.com/tools/godep/Godeps/_workspace/src/github.com/kr/pretty"
 )
@@ -13,11 +14,23 @@ func debugln(a ...interface{}) (int, error) {
 	return 0, nil
 }
 
+func verboseln(a ...interface{}) {
+	if verbose {
+		log.Println(a...)
+	}
+}
+
 func debugf(format string, a ...interface{}) (int, error) {
 	if debug {
 		return fmt.Printf(format, a...)
 	}
 	return 0, nil
+}
+
+func verbosef(format string, a ...interface{}) {
+	if verbose {
+		log.Printf(format, a...)
+	}
 }
 
 func pp(a ...interface{}) (int, error) {