+++ util.go
@@ -10,14 +10,19 @@ import (
 // Stdout, stderr, and the environment are inherited
 // from the current process.
 func runIn(dir, name string, args ...string) error {
+	_, err := runInWithOutput(dir, name, args...)
+	return err
+}
+
+func runInWithOutput(dir, name string, args ...string) (string, error) {
 	c := exec.Command(name, args...)
 	c.Dir = dir
-	output, err := c.CombinedOutput()
+	o, err := c.CombinedOutput()
 
-	if verbose {
+	if debug {
 		fmt.Printf("execute: %+v\n", c)
-		fmt.Printf(" output: %s\n", output)
+		fmt.Printf(" output: %s\n", string(o))
 	}
 
-	return err
+	return string(o), err
 }